Improve how Connected() works to handle opened then closed connections.
Previously, Connected() just checked if the underlying connection was nil.  However, this was wrong for connections that were opened and then closed; the underlying connection was not nil, it still stored data from the previously established connection.  Now, Connected() tried to ping the database which fails if the database connection isn't currently established.

Add tests for Connect(), Close(), and Connected() to handle this bug/error.