Revise how we check for valid/supported database types.
Previously we didn't really check if a db type was valid besides when running the validate() func as part of Connect().  Now, we check that the db type is valid when the NewConfig() func is called.  This just catches errors sooner.

This required bumping go version to 1.18 since we use the experimental generic package slices to handle comparing a db type to the slice/list of valid/supported db types.  This allows us to not have to define our own func that loops through valid db types in a for loop comparing to db type value that was provided.