Starting rewrite of deploy and update schema funcs to not close db connection upon success.
Closing upon success presents an issue when using sqlite in memory db for tests.  In tests, using in memory db is nice since it is very clean.  However, when schema is deployed, currently, the connection is close.  Therefore test has to call Connect().  This then connects to a new in memory db with no db deployed and thus queries cannot be run.

Solutions:
- Deploy and Update don't close db upon success.
- Deploy and Update check if sqlite in-memory db paths are used, and if yes, doesn't close db.
- Deploy and Update take new bool param "closeConnectionAfterSuccess" that dictates if connection is closed.

So far, I think just leaving the connection open and updating docs/comments to say "call Close() if needed after Deploy or Update" and "connection will remain open after Deploy or Update" is the best way.  However, we still need to call Close() when an error gets hit in Deploy or Update just to clean up the connection.