Better support for in-memory sqlite databases by allowing a db connection to remain open after deploying or updating schema.  Clean up some logging.
Previously, deploying or updating a db schema closed the connection after deploy or update completed successfully.  However, this didn't work for in-memory sqlite db since each connection using a new db.  Therefore, when a connection was reestablished after deploy, this new connection didn't see a deployed db, it saw an empty db.  To fix this, deploying or updating a db can now leave the connection open for reuse in running queries.