Rewrite handling of SQLite PRAGMAs to allow more PRAGMAs besides journal_mode to be set.
This was initially needed to support setting busy_timeout for modernc library to match value set by default for mattn library.  By default, mattn library sets busy_timeout to 5000ms.  In order to treat both libraries more closely interchangeable, we need to set the busy_timeout for modernc.  Defining PRAGMAs and running the after connecting to database is just a nice way to handle this and allows for expansion in the future.

Also set the max open connections to 1 by default for SQLite since SQLITE_BUSY errors were getting kicked out randomly when using the modernc library.  Don't know why this was only occurring with modernc but setting this solves the issue.  Note that this really only happened for DELETE journal_mode.

This is a major version change since the config struct changes its definition.