Write new and rewrite existing tests.  Much more test coverage.
Only things that are missing are logging lines, some errors due to MySQL or MariaDB which we don't test against, or funcs for handling update errors.

Provide defined types for UpdateFuncs and DeployFuncs so that users don't have to type stuff like "[]func() error {...}" when defining the list of funcs for a config.