Remove functionality for skipping INSERT INTO statements when deploying db.  This only worked for DeployQueries, not for DeployFuncs, and this led to confusion.
Instead, you should not provide any query or func that inserts data if you want to skip insersion.  This pushes the requirement to handle if a query/func inserts data back to the parent app sqldb is used in which should have much more knowledge about which queries/funcs insert data anyway.