Start work on rewrite to implement better handling of deploying and updating.
Goal is to improve deploying and updating of schema with changes to related translation functions and error handling. This was inconsistent between deploying and updating.

Also planning on revising how creating of a config is done.  Possibly removing a lot of the funcs to help with this (New...) and (Default...) (and related per-db-type funcs) and just have New() (to return a blank Config{}, even though sqldb.Config{}  accomplishes this too) and GetDefault() which uses the singleton/in-package config.  The package is just a bit messy as it is with support for multiple database types and both singleton and external config storage.