Rewrite deploy and updating of schema to use new/rewritten config, clarify some comments, improve how package-level config is stored.
Update DeploySchema and UpdateSchema funcs to use new ...Translators and ...ErrorHandlers in config.  DeploySchema and UpdateSchema now process their related queries and funcs in nearly the same manner.

UpdateSchema no longer uses a transaction; this didn't really matter since a lot of schema update related queries (ALTER...) have an implicit COMMIT called after anyway so encapsulating a bunch of queries in a transaction didn't do what was expected (no complete roll-back on error).  This also now allows a transaction to be created and used within each UpdateQuery as needed; previously this didn't work because you cannot start a transaction inside of a transaction!

New function to create and store config in package. Call New() to get a near-blank config, set fields as needed, then call sqldb.Use() to store the config in the package level variable, as is done when following the singleton style.  This is just a bit cleaner and clearer.